pyproject.toml
# Copyright (c) 2025 - 2026 IQM Finland Oy
# All rights reserved.
#
# Licensed under the Apache License v2.0 with LLVM Exceptions (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
# https://github.com/iqm-finland/QDMI-on-IQM/blob/main/LICENSE.md
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
# WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
# License for the specific language governing permissions and limitations under
# the License.
#
# SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception

[build-system]
requires = ["scikit-build-core>=0.11.6"]
build-backend = "scikit_build_core.build"

[project]
name = "iqm-qdmi"
version = "1.0.0"
requires-python = ">=3.10"
dependencies = [
  "mqt-core[qiskit]==3.5.1",
]

description = "Implementation of a QDMI device for IQM Quantum Computers"
readme = "README.md"
authors = [
  { name = "IQM Finland Oy", email = "developers@meetiqm.com" },
]

license = "Apache-2.0 WITH LLVM-exception"
license-files = ["LICENSE.md"]

classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Intended Audience :: Science/Research",
  "Natural Language :: English",
  "Operating System :: POSIX :: Linux",
  "Programming Language :: C++",
  "Programming Language :: Python",
  "Typing :: Typed",
]

[project.urls]
Homepage = "https://github.com/iqm-finland/QDMI-on-IQM"
Issues = "https://github.com/iqm-finland/QDMI-on-IQM/issues"
Discussions = "https://github.com/iqm-finland/QDMI-on-IQM/discussions"

[project.scripts]
iqm-qdmi = "iqm.qdmi.__main__:main"

[dependency-groups]
test = [
  "pytest>=9.0.1",
  "pytest-console-scripts>=1.4.1",
  "pytest-cov>=7.0.0",
  "pytest-sugar>=1.1.1",
  "pytest-xdist>=3.8.0",
]
docs = [
  "breathe>=4.36.0",
  "furo>=2025.9.25",
  "myst-nb>=1.3.0",
  "sphinx-autoapi>=3.6.1",
  "sphinx-copybutton>=0.5.2",
  "sphinx-design>=0.6.1",
  "sphinx>=8.1.3",
]
dev = [
  {include-group = "test"},
  "nox>=2026.4.10",
  "ty==0.0.29",
]


[tool.scikit-build]
minimum-version = "build-system.requires"
cmake.version = "CMakeLists.txt"
ninja.make-fallback = true
build-dir = "build/{wheel_tag}/{build_type}"
build.targets = ["iqm-qdmi-device"]
wheel.py-api = "py310"
wheel.expand-macos-universal-tags = true
wheel.packages = ["python/iqm"]
wheel.install-dir = "iqm/qdmi/data"
install.components = ["iqm-qdmi-device_Runtime", "iqm-qdmi-device_Development"]

sdist.include = ["python/iqm/qdmi/_version.py"]
sdist.exclude = [
  "**/docs",
  "**/test",
  "sitecustomize.py",
  ".*",
]

[tool.scikit-build.cmake.define]
BUILD_IQM_QDMI_TESTS = "OFF"

[[tool.scikit-build.generate]]
path = "iqm/qdmi/_version.py"
template = '''
version = "${version}"
'''


[tool.pytest]
minversion = "9.0"
strict = true
addopts = [
  "-ra",
  "--numprocesses=auto", # Automatically use all available CPU cores for parallel testing
]
filterwarnings = [
  "error",
]
log_level = "INFO"
testpaths = ["test/python"]


[tool.coverage]
run.source = ["iqm.qdmi"]
run.disable_warnings = [
  "no-sysmon",
]
report.exclude_also = [
  '\.\.\.',
  'if TYPE_CHECKING:',
  'raise AssertionError',
  'raise NotImplementedError',
  'def __dir__()',  # Ignore __dir__ method that exists mainly for better IDE support
  '@overload'  # Overloads are only for static typing
]


[tool.ruff]
line-length = 120
namespace-packages = ["iqm"]
preview = true
fix = true
unsafe-fixes = true
show-fixes = true
src = ["python"]

[tool.ruff.lint]
select = ["ALL"]
ignore = [
  "C90",     # <...> too complex
  "COM812",  # Conflicts with formatter
  "PLR09",   # Too many <...>
  "PLR2004", # Magic value used in comparison
  "S101",    # Use of assert detected
  "S404",    # `subprocess` module is possibly insecure
  "TID252"   # Prefer absolute imports over relative imports from parent modules
]
future-annotations = true

[tool.ruff.lint.isort]
known-first-party = ["iqm.qdmi"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"python/iqm/qdmi/__init__.py" = ["RUF067"]
"python/iqm/qdmi/__main__.py" = ["T201"]
"test/python/**" = ["T20", "INP001"]
"docs/**" = ["T20", "INP001"]


[tool.typos]
default.extend-ignore-re = [
  "(?Rm)^.*(#|//)\\s*spellchecker:disable-line$", # ignore line
  "(?s)(#|//)\\s*spellchecker:off.*?\\n\\s*(#|//)\\s*spellchecker:on", # ignore block
  "FoMaC"
]

[tool.typos.default.extend-words]
optin = "optin"
recal = "recal"
delocate = "delocate"

[tool.cibuildwheel]
build = "cp3*"
archs = "auto64"
test-groups = ["test"]
test-sources = ["test/python"]
test-command = "pytest test/python"
build-frontend = "uv"

[tool.cibuildwheel.linux]
before-all = "yum install -y curl-devel"

[tool.cibuildwheel.macos]
environment = { MACOSX_DEPLOYMENT_TARGET = "11.0" }

[tool.cibuildwheel.windows]
before-build = "uv pip install delvewheel>=1.12.0"
repair-wheel-command = "delvewheel repair -w {dest_dir} {wheel} --namespace-pkg iqm"

[[tool.cibuildwheel.overrides]]
select = "*-musllinux*"
before-all = "apk add curl-dev"


[tool.uv]
cache-keys = [
    { file = "pyproject.toml" },
    { git = { commit = true, tags = true } },
    { file = "src/**/*" },
]
